\name{MilchBig}
\docType{data}
\alias{MilchBig}
\title{Milk components of 50 mammals}
\description{
This dataset contains the fat, protein, lactose and ash of different
milks.
}
\author{
Georg Ohmayer <georg.ohmayer@hswt.de>, \cr
Herbert Seiler 
}
\references{
Ohmayer G., Seiler H. (1985): Numerische Gruppierung und graphische Darstellung von Daten: Ein Methodenvergleich; 
EDV in Medizin und Biologie, Vol. 2, S. 65-73 
}
\usage{MilchBig}
\keyword{datasets}