\name{MilchSmall}
\docType{data}
\alias{MilchSmall}
\title{Milk components of mammals}
\description{
This dataset contains the fat, protein, lactose and ash of different milks. This is a reduced version 
(20 instead of 50 observations).
}
\author{
Georg Ohmayer <georg.ohmayer@hswt.de>, \cr
Herbert Seiler 
}
\references{
Ohmayer G., Seiler H. (1985): Numerische Gruppierung und graphische Darstellung von Daten: Ein Methodenvergleich; 
EDV in Medizin und Biologie, Vol. 2, S. 65-73 
}
\usage{MilchSmall}
\keyword{datasets}